rethink validation, use contract-based approach
